requirements.txt
fastapi
uvicorn
python-dotenv
ollama
elastic-apm
elasticsearch==8.10.1
aiofiles